update(opt): refine series per chart
