+++ opts/series_line.go
@@ -57,3 +57,28 @@ type LineChart struct {
 	// Which will configure the turn point of step line.
 	Step interface{}
 }
+
+// LineData
+// https://echarts.apache.org/en/option.html#series-line.data
+type LineData struct {
+	// Name of data item.
+	Name string `json:"name,omitempty"`
+
+	// Value of a single data item.
+	Value interface{} `json:"value,omitempty"`
+
+	// Symbol of single data.
+	// Icon types provided by ECharts includes 'circle', 'rect', 'roundRect', 'triangle', 'diamond', 'pin', 'arrow', 'none'
+	// It can be set to an image with 'image://url' , in which URL is the link to an image, or dataURI of an image.
+	Symbol string `json:"symbol,omitempty"`
+
+	// single data symbol size. It can be set to single numbers like 10, or
+	// use an array to represent width and height. For example, [20, 10] means symbol width is 20, and height is10
+	SymbolSize int `json:"symbolSize,omitempty"`
+
+	// Index of x axis to combine with, which is useful for multiple x axes in one chart.
+	XAxisIndex int `json:"XAxisIndex,omitempty"`
+
+	// Index of y axis to combine with, which is useful for multiple y axes in one chart.
+	YAxisIndex int `json:"YAxisIndex,omitempty"`
+}
